Move libvirt-specific types to cmd
